pyproject.toml
[project]
name = "llm-ops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.6"
dependencies = [
    "accelerate>=1.10.1",
    "bitsandbytes>=0.48.1",
    "chainlit>=1.0.0",
    "datasets>=4.2.0",
    "gcsfs>=2025.9.0",
    "google-cloud-aiplatform>=1.120.0",
    "google-cloud-bigquery>=3.38.0",
    "google-cloud-storage>=2.19.0",
    "kfp>=2.14.6",
    "pandas>=2.3.3",
    "peft==0.13.2",
    "python-dotenv>=1.1.1",
    "ragas>=0.3.7",
    "requests>=2.31.0",
    "tensorboard>=2.20.0",
    "torch>=2.9.0",
    "transformers==4.46.*",
    "trl>=0.17.0",
]
